refactor: KEEP-237 exclude admin test endpoints from production bundle
Rename app/api/admin/test/**/route.ts to route.staging.ts and gate
compilation via pageExtensions + INCLUDE_TEST_ENDPOINTS build arg.

- next.config.ts: conditional pageExtensions (excludes staging.ts in prod)
- Dockerfile: ARG/ENV passthrough for INCLUDE_TEST_ENDPOINTS
- docker-bake.hcl: variable declaration + arg on app/sentry-upload targets
- build-images.yml: set INCLUDE_TEST_ENDPOINTS=true for non-prod refs
- Tests updated to import from route.staging

Prod build produces zero admin/test routes in .next; staging/PR builds
still compile them. Runtime gate in lib/admin-auth.ts retained as
defense-in-depth.
