+++ tests/integration/admin-invitation-route.test.ts
@@ -34,7 +34,7 @@ vi.mock("@/lib/db/schema", () => ({
   },
 }));
 
-import { GET } from "@/app/api/admin/test/invitation/route";
+import { GET } from "@/app/api/admin/test/invitation/route.staging";
 
 function createRequest(email?: string, token?: string): Request {
   const url = new URL("http://localhost:3000/api/admin/test/invitation");
